weather-etl-pipeline/requirements.txt
requests==2.31.0
boto3==1.26.174
python-dotenv==1.0.0
